refactor(map): unify color key with contacts staleness key
Map and Contacts pages now share StalenessColorKey, which uses the
contacts copy ("What the colors mean" + per-bucket criteria) and adds
the existing color-picker edit affordance to both popovers. Map legend
trigger swapped to faCircleInfo so the icon reads as informational
rather than a layer stack.
